Cleaned up Shotgun, changed type of getPelletCount from double to int.

+++ src/main/java/me/KiwiLetsPlay/KiwiField/weapon/heavy/Nova.java
@@ -53,7 +53,7 @@ public class Nova implements Shotgun {
 	}
 	
 	@Override
-	public double getPelletCount() {
+	public int getPelletCount() {
 		return 8;
 	}
